Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.2.1",
    "fastapi>=0.115.8",
    "pyjwt>=2.10.1",
    "pymongo>=4.11",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
]
